Codenarc: Adjusting rules for Ratpack
I had copied this from Bloom's main repo, which had a bunch of Grails
stuff in it. Further, I adjusted some of the rules to be in line with
issues we normally need to suppress. I think it'd be good to address
some of the AbcMetric stuff, but at the moment this seems fine.