Added a remember me token.
A cookie is written containing the username and a secret hash. When
the cookie is read later, if the secret hash matches what would have
been expected for that user, then the user is automatically logged in.

Uses BCrypt to hash to cookies just like we do for passwords.

This is used for username/password logins but not for twitter
logins.